if the num of motors > 2, name should be plural
